Changing project structure to ROS standard and fixing some warnings

+++ src/test_throttler.cc
@@ -1,5 +1,5 @@
 #include <ros/ros.h>
-#include "BagThrottler.h"
+#include "rc_bagthrottler/BagThrottler.h"
 
 #include <sensor_msgs/JointState.h>
 #include <message_filters/subscriber.h>
@@ -14,24 +14,30 @@ using namespace std;
 void fastCallback(JointStateConstPtr msg)
 {
   static int lastSeq = -1;
-  if (lastSeq>0 && lastSeq+1!=msg->header.seq) {
+
+  if (lastSeq>0 && lastSeq+1!=static_cast<int>(msg->header.seq))
+  {
     throw runtime_error("!!!fastCallback missed event!!!");
   }
+
   lastSeq = msg->header.seq;
   ROS_INFO_STREAM("fastCallback with low computation, stamp: " <<
-       msg->header.stamp << " seq: " << msg->header.seq);
+                  msg->header.stamp << " seq: " << msg->header.seq);
 }
 
 ros::Publisher pub;
 void slowCallback(JointStateConstPtr msg)
 {
   static int lastSeq = -1;
-  if (lastSeq>0 && lastSeq+1!=msg->header.seq) {
+
+  if (lastSeq>0 && lastSeq+1!=static_cast<int>(msg->header.seq))
+  {
     throw runtime_error("!!!slowCallback missed event!!!");
   }
+
   lastSeq = msg->header.seq;
   ROS_INFO_STREAM("slowCallback with heavy computation, stamp: " <<
-       msg->header.stamp << " seq: " << msg->header.seq);
+                  msg->header.stamp << " seq: " << msg->header.seq);
   sleep(1);
 
   JointStatePtr outmsg(new JointState(*msg));
@@ -41,16 +47,19 @@ void slowCallback(JointStateConstPtr msg)
 void syncCallback(JointStateConstPtr msg)
 {
   ROS_INFO_STREAM("syncCallback with heavy computation, stamp: " <<
-                                                                 msg->header.stamp << " seq: " << msg->header.seq);
+                  msg->header.stamp << " seq: " << msg->header.seq);
   static int lastSeq = -1;
-  if (lastSeq>0 && lastSeq+1!=msg->header.seq) {
+
+  if (lastSeq>0 && lastSeq+1!=static_cast<int>(msg->header.seq))
+  {
     throw runtime_error("!!!syncCallback missed event!!!");
   }
+
   lastSeq = msg->header.seq;
   sleep(1);
 
   JointStatePtr outmsg(new JointState(*msg));
-//  pub.publish(outmsg);
+  //  pub.publish(outmsg);
 }
 
 /*******************************************************************************
@@ -63,13 +72,13 @@ int main(int argc, char *argv[])
   ros::init(argc, argv, "first_node");
 
   // second, we configure bag throttler to a specific rosbag topic
-//  BagThrottler::throttle<JointState>("/iiwa/joint_states");
+  //  BagThrottler::throttle<JointState>("/iiwa/joint_states");
 
   // finally, we can register our actual subscribers
   ros::NodeHandle nh;
-//  pub = nh.advertise<JointState>("/joint_states", 1000);
-//  ros::Subscriber sub = nh.subscribe("/iiwa/joint_states", 5, &fastCallback);
-//  ros::Subscriber sub2 = nh.subscribe("/iiwa/joint_states", 2, &slowCallback);
+  //  pub = nh.advertise<JointState>("/joint_states", 1000);
+  //  ros::Subscriber sub = nh.subscribe("/iiwa/joint_states", 5, &fastCallback);
+  //  ros::Subscriber sub2 = nh.subscribe("/iiwa/joint_states", 2, &slowCallback);
 
   message_filters::Subscriber<JointState> joint_sub(nh, "/iiwa/joint_states", 2);
   message_filters::Subscriber<JointState> joint_sub2(nh, "/iiwa/joint_states", 10);
